+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog Fab-manager
 
 - Fix a bug: incorrect date range in statistics
+- Fix a bug: hide project categories filter if there is no project categories
+- Improvement : dialog confirm to remove a project category
 
 ## v6.0.10 2023 July 13